moved terminateeventdefinition
